pyproject.toml
[project]
name = "payment-system-maryia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
]
